Updated implementation of Board, Piece and Player. Add the scanner and GameView implementation
Implements test and code of toString in Piece, Finished movePiece, chooseToEnterPiece, displayPieces and choosePiece in Player. Add the implementation of Scanner to get the inputs of the players.Add the GameView to print the steps of the game. Firts implementation of getNextSquare and displayBoard on Board, the transition from global path to final square does not work correctly.